Folder structure updated
