+++ src/main/java/com/controller/EmployeeLogout.java
@@ -1,0 +1,19 @@
+package com.controller;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+import java.io.IOException;
+
+@WebServlet("/EmployeeLogout")
+public class EmployeeLogout extends HttpServlet {
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        HttpSession session = request.getSession();
+        session.removeAttribute("username");
+        session.invalidate();
+        response.sendRedirect("employee.jsp");
+    }
+}
