+++ src/main/java/com/interfaces/EmployeeDatabase.java
@@ -1,0 +1,5 @@
+package com.interfaces;
+
+public interface EmployeeDatabase {
+    public boolean checkLogin(String username, String password);
+}
